The app can change the speed rate when the configuration is updated.
