+++ app/src/main/java/ar/com/westsoft/listening/screen/dictationgame/game/GameConsoleViewModel.kt
@@ -1,0 +1,35 @@
+package ar.com.westsoft.listening.screen.keyboard.ar.com.westsoft.listening.screen.dictationgame.game
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import ar.com.westsoft.listening.domain.dictationgame.engine.GetDictationGameStateFlowUseCase
+import ar.com.westsoft.listening.domain.dictationgame.engine.MoveToParagraphUseCase
+import ar.com.westsoft.listening.domain.dictationgame.engine.SpeakOutUseCase
+import ar.com.westsoft.listening.screen.dictationgame.game.DictGameState
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
+import javax.inject.Inject
+
+@HiltViewModel
+class GameConsoleViewModel @Inject constructor(
+    getDictationGameStateFlowUseCase: GetDictationGameStateFlowUseCase,
+    private val speakOutUseCase: SpeakOutUseCase,
+    private val moveToParagraphUseCase: MoveToParagraphUseCase
+) : ViewModel() {
+
+    val dictationGameStateFlow: StateFlow<DictGameState> =
+        getDictationGameStateFlowUseCase(viewModelScope)
+
+    fun onLetterClicked(offset: Int) {
+        viewModelScope.launch {
+            speakOutUseCase(offset = offset)
+        }
+    }
+
+    fun onParagraphClick(paragraphIdx: Int) {
+        viewModelScope.launch {
+            moveToParagraphUseCase(paragraphIdx)
+        }
+    }
+}
