+++ app/src/main/java/ar/com/westsoft/listening/screen/dictationgame/navigation/GameCreationGameStatus.kt
@@ -2,7 +2,7 @@ package ar.com.westsoft.listening.screen.dictationgame.navigation
 
 sealed class GameCreationGameStatus {
     object Uninitialized : GameCreationGameStatus()
-    class Completed(val gui: Long) : GameCreationGameStatus()
+    object Completed : GameCreationGameStatus()
     object Error : GameCreationGameStatus()
     object IsDownloading: GameCreationGameStatus()
 }
