+++ app/src/main/java/ar/com/westsoft/listening/screen/dictationgame/navigation/SelectDictationGameViewModel.kt
@@ -3,8 +3,11 @@ package ar.com.westsoft.listening.screen.dictationgame.navigation
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import ar.com.westsoft.listening.data.engine.DictationGameHeader
+import ar.com.westsoft.listening.di.DefaultDispatcher
+import ar.com.westsoft.listening.domain.dictationgame.engine.SetupDictationUseCase
 import ar.com.westsoft.listening.domain.dictationgame.repository.GetDictationGameLabels
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
@@ -12,7 +15,9 @@ import javax.inject.Inject
 
 @HiltViewModel
 class SelectDictationGameViewModel @Inject constructor(
-    private val getDictationGameLabels: GetDictationGameLabels
+    private val getDictationGameLabels: GetDictationGameLabels,
+    private val setupDictationGameUseCase: SetupDictationUseCase,
+    @DefaultDispatcher private val defaultDispatcher: CoroutineDispatcher
 ) : ViewModel() {
 
     private val _games = MutableStateFlow<List<DictationGameHeader>>(emptyList())
@@ -21,6 +26,14 @@ class SelectDictationGameViewModel @Inject constructor(
     init{
         viewModelScope.launch {
             _games.emit(getDictationGameLabels())
+        }
+    }
+
+    fun onInitializeProgress(gui: Long) {
+        viewModelScope.launch(defaultDispatcher) {
+            setupDictationGameUseCase(gui)
+            println("**** set game: $gui.")
+            println("**** set new DictationGameFlow.")
         }
     }
 }
